requirements.txt
google-cloud-storage>=2.16.0
google-auth>=2.29.0
requests>=2.32.0